gdb/testsuite: use the correct .debug_str section name for DW_FORM_strp
When handling DWARF attributes of the form DW_FORM_strp the strings
should be placed in the .debug_str section, not .debug_string as they
currently are by the DWARF assembler (in lib/dwarf.exp).

I've added a test.  This is as much to test the DWARF generator as it
is to test GDB as GCC makes frequent use of DW_FORM_strp so we can be
pretty sure this part of GDB is already well tested.

gdb/testsuite/ChangeLog:

	* gdb.dwarf2/dw2-using-debug-str.c: New file.
	* gdb.dwarf2/dw2-using-debug-str.exp: New file.
	* lib/dwarf.exp (Dwarf::DW_FORM_strp): Create .debug_str section,
	not .debug_string.